// frontend/src/app/models/nekretnina.ts
export class Nekretnina{
    idN: number;
    naziv: string;
    tip: string;
    grad: string;
    opstina: string;
    mikrolokacija: string;
    ulica: string;
    kvadratura: number;
    cena: number;
    oglasivac: string;
    godIzgradnje: string;
    stanje: string;
    tipGrejanja: string;
    sprat: number;
    ukupnaSpratnost: number;
    parking: string;
    opis:string;
    karakteristike:string;
    slike:Array<String>;
    linije:string;
    brSoba:number;
    mesecneRezije: number;
    tipOglasivaca: string;
    prodato: string;
    izmenaDatum: string;
    izmenaVreme: string;
    randomSlika: number;
}